phan_quyen

+++ datn/app/Http/Controllers/Admin/VoucherController.php
@@ -17,6 +17,7 @@ class VoucherController extends Controller
 {
     //
     public function index(){
+        $this->authorize('admin');
         $today = Carbon::now('Asia/Ho_Chi_Minh')->format('Y-m-d H:i:s');
         $voucher = Voucher::all();
         $voucher = Voucher::paginate(5);
@@ -24,6 +25,7 @@ class VoucherController extends Controller
 
     }
     public function addVoucherToCart(Request $req){
+        $this->authorize('admin');
         $today = Carbon::now('Asia/Ho_Chi_Minh')->format('Y-m-d H:i:s');
         $voucherCode = Voucher::where('code',$req->voucherCode)->first();
         if(!$voucherCode){
@@ -68,9 +70,11 @@ class VoucherController extends Controller
         }
     }
     public function create_voucher(Request $request){
+        $this->authorize('admin');
         return view('admin.voucher.add-voucher');
     }
     public function saveAdd(Request $request){
+        $this->authorize('admin');
         if($request->isMethod('POST')){
             $rule = [
                 'name' => 'required',
@@ -111,10 +115,12 @@ class VoucherController extends Controller
     }
 
     public function edit_voucher($id ,Request $request){
+        $this->authorize('admin');
         $show = Voucher::find($id);
             return view('admin.voucher.edit-voucher', compact('show'));
         }
     public function update_voucher($id,VoucherRequest $request){
+        $this->authorize('admin');
         $dt_create = Carbon::now('Asia/Ho_Chi_Minh')->toDateTimeString();
         $data= $_POST;
         $model = Voucher::find($id);
@@ -127,7 +133,8 @@ class VoucherController extends Controller
     }
 
     public function destroy($id)
-    {
+    {   
+        $this->authorize('admin');
         $User = Voucher::find($id);
         $User->delete();
         return back();
